Expose concurrency as an option
This also limits the maximum number of concurrent operations to 16. This is
good practice for large files with thousands of parts. Instead of blasting S3
server with all thousands of requests simultaneously, making a small number of
them at a time increases stability.

This is also necessary when you want to create upload stream from the same
source file, because their is a limit to how many simultaneous streams you can
create from a file.

This does not result in any speed loss at all. On the contrary, it increases
throughput for large files.