Replace multiple files with multiple streams created from the same file
This gets rid of the initial delay which is necessary to split the source
file into multiple temporary files. For large files (> 5GB) this delay can
be at the order of minutes which is more than the time required to actually
upload the file.

Of course this limits the number of simultaneous requests we can make.